01_core_system/fixed_system/requirements.txt
# AutoGen e dependências base
pyautogen==0.2.18
openai>=1.3
python-dotenv==1.1.0
requests>=2.31.0
httpx>=0.24.0

# Para modelos locais
llama-cpp-python==0.2.90
ollama>=0.1.7

# Interface e utilidades
rich>=13.0.0
colorama>=0.4.6
tqdm>=4.65.0

# Cache e persistência
diskcache>=5.6.0

# Análise e processamento
numpy==1.26.4
pandas>=2.0.0

# Para fact-checking (opcional)
duckduckgo-search>=4.0.8
rapidfuzz>=3.6.1